# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientPhone
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
    put:
      operationId: updatePrescription
      tags: [prescriptions]
      summary: Update a prescription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/verify/{ref}:
    get:
      operationId: verifyPrescription
      tags: [prescriptions]
      summary: Public verification of a prescription by reference number
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionVerification"
...
  /admin/prescriptions:
    get:
      operationId: listAdminPrescriptions
      tags: [prescriptions]
      summary: Search all prescriptions (admin repository)
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: countryId
          in: query
          schema:
            type: integer
        - name: locationId
          in: query
          schema:
            type: integer
        - name: patientName
          in: query
          schema:
            type: string
        - name: patientPhone
          in: query
          schema:
            type: string
        - name: referenceNo
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Prescription repository results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPrescriptionList"
        "403":
          description: Forbidden

  # ---- AUDIT LOGS ----
...
  /admin/patient-timeline:
    get:
      operationId: getPatientTimeline
      tags: [admin]
      summary: Aggregated timeline for a patient (admin/doctor)
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientTimeline"
        "403":
          description: Forbidden

  # ---- DATA MIGRATION CENTER ----
...
components:
  schemas:
...
    Prescription:
      type: object
      required: [id, doctorId, patientName, createdAt]
      properties:
        id:
          type: integer
        referenceNo:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, final, pending_investigation]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        appointmentId:
          type: ["integer", "null"]
        oldPatient:
          type: boolean
        freePatient:
          type: boolean
        consultationFee:
          type: ["integer", "null"]
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        patientAge:
          type: ["integer", "null"]
        patientGender:
          type: ["string", "null"]
        patientWeight:
          type: ["string", "null"]
        patientHeight:
          type: ["string", "null"]
        chiefComplaint:
          type: ["string", "null"]
        vitals:
          type: ["string", "null"]
        examination:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        investigations:
          type: ["string", "null"]
        advice:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItem"
        createdAt:
          type: string
    PrescriptionItem:
      type: object
      required: [id, medicineName]
      properties:
        id:
          type: integer
        medicineId:
          type: ["integer", "null"]
        medicineName:
          type: string
        genericName:
          type: ["string", "null"]
        strength:
          type: ["string", "null"]
        dosageForm:
          type: ["string", "null"]
        dose:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        mealTiming:
          type: ["string", "null"]
        instruction:
          type: ["string", "null"]
    PrescriptionInput:
      type: object
      required: [doctorId, patientName, items]
      properties:
        status:
          type: string
          enum: [draft, final, pending_investigation]
        doctorId:
          type: integer
        appointmentId:
          type: integer
        oldPatient:
          type: boolean
        freePatient:
          type: boolean
        consultationFee:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientAge:
          type: integer
        patientGender:
          type: string
        patientWeight:
          type: string
        patientHeight:
          type: string
        chiefComplaint:
          type: string
        vitals:
          type: string
        examination:
          type: string
        diagnosis:
          type: string
        investigations:
          type: string
        advice:
          type: string
        followUpDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItemInput"
    PrescriptionItemInput:
      type: object
      required: [medicineName]
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        genericName:
          type: string
        strength:
          type: string
        dosageForm:
          type: string
        dose:
          type: string
        duration:
          type: string
        mealTiming:
          type: string
        instruction:
          type: string
    PrescriptionVerification:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        referenceNo:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
...
    AdminPrescriptionList:
      type: object
      required: [prescriptions, total]
      properties:
        total:
          type: integer
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
    TimelineEvent:
      type: object
      required: [id, type, date]
      properties:
        id:
          type: integer
        type:
          type: string
        date:
          type: string
        title:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        bp:
          type: ["string", "null"]
        pulse:
          type: ["string", "null"]
        temp:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        labReportUrl:
          type: ["string", "null"]
        prescriptionUploadUrl:
          type: ["string", "null"]
        referenceNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    PatientTimeline:
      type: object
      required: [patientPhone, events]
      properties:
        patientPhone:
          type: string
        patientName:
          type: ["string", "null"]
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
...